apps/requirements.txt
# Flask for web apps
flask>=3.0.0

# Environment management
python-dotenv>=1.0.0

# HTTP client
requests>=2.31.0

# Pygame for display apps
pygame>=2.5.0
